requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
cffi==1.12.3
cryptography==2.6.1
Cython==0.29.6
ecdsa==0.13.2
falcon==2.0.0
falcon-autocrud==1.0.36
googleapis-common-protos==1.6.0
grpcio==1.20.1
grpcio-tools==1.20.1
jsonschema==3.0.1
pkg-resources==0.0.0
protobuf==3.7.1
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
pyrsistent==0.15.2
six==1.12.0
SQLAlchemy==1.3.3
uWSGI==2.0.18
